requirements.txt
google-genai
mcp
openai
langchain
jupyter
